Changed Food Price from INT to DOUBLE

+++ DataObjects/Food/AddFoodRequest.cs
@@ -8,7 +8,7 @@ namespace FastFoodAPI.DataObjects.Food
     public class AddFoodRequest
     {
         public required string Name { get; set; }
-        public required int Price { get; set; }
+        public required double Price { get; set; }
         public string Description { get; set; } = "Description was not set";
         public FoodTypes Type { get; set; } = FoodTypes.Burguer;
