+++ DataObjects/Food/AddFoodResponse.cs
@@ -9,7 +9,7 @@ namespace FastFoodAPI.DataObjects.Food
     {
         public int Id { get; set; }
         public required string Name { get; set; }
-        public required int Price { get; set; }
+        public required double Price { get; set; }
         public string Description { get; set; } = "Description was not set";
         public FoodTypes Type { get; set; } = FoodTypes.Burguer;
